pyproject.toml
[project]
name = "reviewbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "fastapi>=0.129.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openrouter>=0.0.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "tortoise-orm>=1.1.5",
    "tzdata>=2025.3",
    "uvicorn>=0.41.0",
    "watchdog>=6.0.0",
]
